Add option to use custom shell
By default `shell-env` uses the login shell to get the environment variables from, this PR adds an optional argument that can be used to override that with a shell of choice.

We need this over at [hyperterm](https://github.com/zeit/hyperterm)

Also, i updated it to use `execa.sync` as stated by the TODO ⭐

Should it throw an error if `shell` is passed and can't be executed?
Instead of silently returning process.env..

Closes #3